Refactor: Add companyGuid parameter to client and debt queries
This commit modifies the `ClientDao` and `ClientRepository` to include the `companyGuid` parameter when fetching client information and client debts.

The `ClientViewModel` and `ClientListViewModel` have been updated to utilize the new `companyGuid` parameter from `SharedViewModel` when retrieving data. This ensures that client data and debts are filtered correctly based on the current company.